+++ components/Home/CreatePost/CreatePost.jsx
@@ -12,14 +12,8 @@ import { newPost } from '@epicapp/services/post';
 ///utils
 import { convertBase64 } from '@epicapp/utils/image';
 
-export default function CreatePost({ author, removeRounded }) {
-  var setRoundedTop = '3xl';
-  if (removeRounded === 'true') {
-    setRoundedTop = 'none'
-  }
-
-//export default function CreatePost({ author }) {
-  //const queryClient = useQueryClient();
+export default function CreatePost({ author }) {
+  const queryClient = useQueryClient();
   const [contentType, setContentType] = useState('text/plain');
   const [imageFileName, setImageFileName] = useState(null);
   const [visibility, setVisibility] = useState({
@@ -85,7 +79,7 @@ export default function CreatePost({ author, removeRounded }) {
   }
 
   return (
-    <section className={`rounded-t-${setRoundedTop} rounded-b-3xl bg-surface p-4`}>
+    <section className="rounded-3xl bg-surface p-4">
       <form onSubmit={formHandler}>
         <div className="flex gap-4">
           <div>
@@ -160,7 +154,9 @@ export default function CreatePost({ author, removeRounded }) {
             </Button>
           </div>
           <div className="w-full">
-            <p className="my-1 text-lg text-textAlt py-2">What's on your mind?</p>
+            <p className="my-1 py-2 text-lg text-textAlt">
+              What's on your mind?
+            </p>
             <div className="w-full overflow-hidden rounded-2xl bg-foreground text-text">
               <input
                 className="h-9 w-full border-b border-layer bg-transparent p-3 placeholder:text-textAlt/20 focus:outline-none"
