+++ pages/details.js
@@ -4,40 +4,39 @@ import { useQuery } from 'react-query';
 //components
 import HomeLayout from '@epicapp/layouts/HomeLayout';
 import Profile from '@epicapp/components/Home/Profile';
-import CreatePost from '@epicapp/components/Home/CreatePost';
-import Stream from '@epicapp/components/Home/Stream';
 
 //services
 import { getAuthorDetails } from '@epicapp/services/author';
 
-export default function Homepage() {
+export default function ProfilePage() {
   const author = useQuery(['author'], getAuthorDetails, {
-    staleTime: Infinity,
+    retry: 1,
+    staleTime: 10000,
   });
 
   return (
     <>
       <Head>
-        <title>Epic App</title>
+        <title>Profile</title>
         <meta name="description" content="Generated by create next app" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <main className="h-screen w-screen">
+      <main className="h-screen w-screen overflow-y-auto bg-background">
         {author.isLoading ? (
           <div className="flex h-full items-center justify-center text-9xl text-primary">
             {/* // maybe a ekelton loading animation here? */}
             <i className="fa-solid fa-spinner-third animate-spin bg-transparent text-2xl text-primary" />
           </div>
         ) : (
-          <HomeLayout route="HOME" author={author.data?.data}>
-            <div className="grid grid-cols-8 gap-8">
-              <section className="col-span-2 w-full">
+          <HomeLayout route="PROFILE" author={author.data?.data}>
+            <div className="grid grid-cols-4 gap-8">
+              <section className="col-span-2 col-start-2 w-full">
+                {/* might need to make another component, idk depending on what you want to do */}
                 <Profile author={author.data?.data} />
               </section>
               <section className="col-span-4 flex flex-col gap-6">
-                <CreatePost author={author.data?.data} />
-                <Stream author={author.data?.data} />
+                {/* SHOULD ONLY HAVE POSTS FROM THIS AUTHOR */}
               </section>
             </div>
           </HomeLayout>
@@ -47,14 +46,13 @@ export default function Homepage() {
   );
 }
 
-//TODO: DO THIS WAY OR HAVE PUBLIC VIEW
 // export async function getServerSideProps(context) {
 //   if (!context.req.headers.cookie?.includes('access=')) {
 //     return {
 //       props: {},
 //       redirect: {
 //         permanent: false,
-//         destination: '/login',
+//         destination: '/',
 //       },
 //     };
 //   }
